requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.3
mysql-connector-python==8.0.24
protobuf==3.15.8
PyMySQL==1.0.2
six==1.15.0
SQLAlchemy==1.4.11
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
